requirements.txt
fastapi
uvicorn
pydantic
aiosqlite
sqlalchemy
python-jose
passlib